requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-binance>=1.0.19
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=20.1.0